+++ lib/post/pages/post_page.dart
@@ -493,6 +493,9 @@ class _PostPageState extends State<PostPage> {
     AuthBloc authBloc = context.read<AuthBloc>();
     AccountBloc accountBloc = context.read<AccountBloc>();
 
+    final ThunderState state = context.read<ThunderBloc>().state;
+    final bool reduceAnimations = state.reduceAnimations;
+
     if (!authBloc.state.isLoggedIn) {
       showSnackbar(context, AppLocalizations.of(context)!.mustBeLoggedInComment);
     } else {
@@ -513,6 +516,7 @@ class _PostPageState extends State<PostPage> {
       Navigator.of(context)
           .push(
         SwipeablePageRoute(
+          transitionDuration: reduceAnimations ? const Duration(milliseconds: 100) : null,
           canOnlySwipeFromEdge: true,
           backGestureDetectionWidth: 45,
           builder: (context) {