+++ lib/post/widgets/comment_view.dart
@@ -96,6 +96,8 @@ class _CommentSubviewState extends State<CommentSubview> with SingleTickerProvid
     final theme = Theme.of(context);
     final ThunderState state = context.read<ThunderBloc>().state;
 
+    final reduceAnimations = state.reduceAnimations;
+
     if (!widget.viewFullCommentsRefreshing && _removeViewFullCommentsButton) {
       _animatingIn = true;
       _fullCommentsAnimation.reverse();
@@ -119,6 +121,7 @@ class _CommentSubviewState extends State<CommentSubview> with SingleTickerProvid
         }
       },
       child: ScrollablePositionedList.builder(
+        physics: reduceAnimations ? const BouncingScrollPhysics() : null,
         addSemanticIndexes: false,
         itemScrollController: widget.itemScrollController,
         itemPositionsListener: widget.itemPositionsListener,