+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Add option to disabling graying out read posts - contribution from @micahmo
 - Show sort type icon - contribution from @micahmo
 - Downvote actions will be disabled when instances have downvotes disabled
+- Added accessibility settings to reduce animations/motion
 - Automatically save drafts for posts and comments - contribution from @micahmo
 - Newly created comments get inserted into comment list correctly without losing your scroll position. If comment is top level, the list scrolls to your comment. The comment also gets highlighted - contribution from @ajsosa